Abstracted rows and columns
