+++ webpack/scenes/Packages/PackagesPage.js
@@ -1,8 +1,7 @@
 import React, { Component } from "react"
 import { Grid, Row, Col, Table } from 'patternfly-react';
-
-const headerFormat = value => <th>{value}</th>;
-const cellFormat = value => <td>{value}</td>;
+import packagesColumns from './columns';
+import mockRows from './mockRows';
 
 class PackagesPage extends Component {
   constructor(props) {
@@ -24,30 +23,10 @@ class PackagesPage extends Component {
                 striped
                 bordered
                 hover
-                columns={[
-                  {header: {label: 'First Name',formatters: [headerFormat]},cell: {formatters: [cellFormat]},property: 'first_name'},
-                  {header: {label: 'Last Name',formatters: [headerFormat]},cell: {formatters: [cellFormat]},property: 'last_name'},
-                  {header: {label: 'Username',formatters: [headerFormat]},cell: {formatters: [cellFormat]},property: 'username'},
-                ]}
+                columns={packagesColumns}
               >
               <Table.Header/>
-              <Table.Body rows={[
-                  {
-                    id: 0,
-                    first_name: 'Dan',
-                    last_name: 'Abramov',
-                  },
-                  {
-                    id: 1,
-                    first_name: 'Sebastian',
-                    last_name: 'Markbåge',
-                  },
-                  {
-                    id: 2,
-                    first_name: 'Sophie',
-                    last_name: 'Alpert',
-                  },
-                ]} rowKey="id" />
+              <Table.Body rows={mockRows} rowKey="id" />
               </Table.PfProvider>
             </Col>
           </Row>
