+++ webpack/scenes/Packages/mockRows.js
@@ -1,0 +1,19 @@
+const mockRows = [
+   {
+     id: 0,
+     first_name: 'Dan',
+     last_name: 'Abramov',
+   },
+   {
+     id: 1,
+     first_name: 'Sebastian',
+     last_name: 'Markbåge',
+   },
+   {
+     id: 2,
+     first_name: 'Sophie',
+     last_name: 'Alpert',
+   },
+]
+
+export default mockRows;
